Altering (mostly by making internal data Dwarf_Unsigned) so that we avoid many warnings. 	modified:   src/lib/libdwarf/dwarf_opaque.h 	modified:   src/lib/libdwarf/dwarf_query.c 	modified:   src/lib/libdwarf/dwarf_rnglists.c 	modified:   src/lib/libdwarf/dwarf_setup_sections.c 	modified:   src/lib/libdwarf/dwarf_str_offsets.c 	modified:   src/lib/libdwarf/dwarf_util.c
Fixed a mistaken return-value type from int to Dwarf_Unsigned.
See Changes in doc/libdwarf.dox
	modified:   src/lib/libdwarf/dwarf_query.c
	modified:   doc/libdwarf.dox
	modified:   src/lib/libdwarf/libdwarf.h

+++ src/lib/libdwarf/dwarf_setup_sections.c
@@ -137,7 +137,7 @@ set_up_section(Dwarf_Debug dbg,
         name we care about as possibly compressed, which
         is to say bigger than any standard section name. */
 #define SECNAMEMAX 30
-    int secnamelen = strlen(secname);
+    size_t secnamelen = strlen(secname);
     /* static const char *dprefix = ".debug_"; */
 #define DPREFIXLEN 7
     static const char *zprefix = ".zdebug_";
